Yuuta

+++ GameSources/BreakWall.cpp
@@ -64,7 +64,7 @@ namespace basecross {
 		if (m_Hp <= 0)
 		{	
 			int damage = -20;//これを壊したときの患者へのダメージ量
-			GetStage()->GetSharedGameObject<StageManager>(L"StageManager")->SetHp(damage);//ダメージを与える
+			GetStage()->GetSharedGameObject<StageManager>(L"StageManager")->SetHp((float)damage);//ダメージを与える
 			int piece = 0;//どれくらいピースを手に入るかを決める
 			piece = rand() % 3;//ランダムにどのピースが出るのかを決める変数
 
@@ -82,7 +82,7 @@ namespace basecross {
 			m_ptrCollider->SetFixed(false);//これでぶつかっても動かないようにする
 
 		}
-		m_Trans->SetPosition(m_Position);//位置を設定	
+		m_Trans->SetPosition(m_Position);//位置を設定
 
 
 	};
@@ -113,8 +113,19 @@ namespace basecross {
 		if (Other->FindTag(L"Player"))
 		{
 			m_Position = m_StartPosition;
+			//m_ptrCollider->SetFixed(true);//これでぶつかっても動かないようにする
 		}
 	}
+
+	//void BreakWall::OnCollisionExit(shared_ptr<GameObject>& Other)
+	//{
+	//	if (Other->FindTag(L"Player"))
+	//	{
+	//		//m_Position = m_StartPosition;
+	//		m_ptrCollider->SetFixed(false);//これでぶつかっても動かないようにする
+	//	}
+
+	//}
 
 }
 //end namespace basecross
